feat(nvs_structures): make nvs_set_sgp30_baseline static
- Changed `nvs_set_sgp30_baseline` function to static to limit its scope to the file.
- Added `storage_set_thingsboard_cfg` function to set ThingsBoard configuration.

fix(sgp30): update sgp30_start_measuring parameter type

- Changed the parameter type of `sgp30_start_measuring` from `uint64_t` to `uint32_t` for better compatibility.

refactor(main): remove unused event handler and fix function names

- Removed `new_send_time_event_handler` as it was no longer used.
- Renamed `sgp30_on_new_interval` to `mqtt_on_new_interval` for better clarity.
- Fixed incorrect function calls and formatting issues in `app_main`.

+++ components/sgp30/include/sgp30.h
@@ -183,7 +183,7 @@ sgp30_init(esp_event_loop_handle_t loop, sgp30_measurement_t *baseline);
  *  - ESP_OK : if the timer was started successfully
  *  - ESP_FAIL : if the timer could not be started
  */
-esp_err_t sgp30_start_measuring(uint64_t s);
+esp_err_t sgp30_start_measuring(uint32_t s);
 /**
  * @brief Restarts the measurement timer with a new interval.
  * This function restarts the measurement timer with a new interval.
